AccountInfo Completed
The AccountInfo page has been setup thusly that it allows for the input of model data.

Changed a number of files, including the files in the Logic, as they were not fit to be there.

To make an ammend to the previous statements, Logic has NOT been implemented yet, only the ability to display model data on the pages.
Logic should be implemented and handling all the hard coded data samples and the controllers should only function to use the logic interfaces to get this hard coded data.