Bestiary Finished
The bestiary now functions as it should with data being supplied to the page by the controller.
The page has also been fixed up by not selecting an active enemy initially, instead it selects nothing and the user gets to select one.

Fixed a few issues with the Register page.