Leaderboard Logic Finished + Fixes
Implemented the logic for the Leaderboard page which now allows the controllers to receive their data from the logic layer instead of their own hard coded values.

Implemented Session to allow for users to be logged in and navigate accordingly.
Fixed a few issues that occured when implementing the use of Session.