DAL Foundation Lain
The foundation for the DAL, which will be filled upon DAL implementation, has been added to the project and ready for use.

Do note that the Logic still uses hard coded data which needs to be replaced.
And the Logic still needs to get the references to the correct DAL components. (the interfaces)