Altered Logic and View
Changed the Logic and View to be more correct according to Single responsibility.
The obtaining and keeping of interface collections has been moved to special Collection classes (UserCollection, EnemyCollection, Leaderboard) as they do not fall under the responsibility of an individual element.

Fixed a number of issues in the view that occured when transfering from the old to the new classes.